sign conversion is fun

+++ unit_tests/test_itoa_rev.cpp
@@ -6,12 +6,12 @@
 
 TEST_GROUP(itoa) { char buf[64]; };
 
-#define NPF_ITOA_CHECK(EXPECT_STR, VAL)             \
-    do {                                            \
-        int const n = npf__itoa_rev(buf, VAL);      \
-        CHECK_EQUAL(strlen(EXPECT_STR), (size_t)n); \
-        buf[n] = 0;                                 \
-        STRCMP_EQUAL(EXPECT_STR, buf);              \
+#define NPF_ITOA_CHECK(EXPECT_STR, VAL)          \
+    do {                                         \
+        int const n = npf__itoa_rev(buf, VAL);   \
+        CHECK_EQUAL((int)strlen(EXPECT_STR), n); \
+        buf[n] = 0;                              \
+        STRCMP_EQUAL(EXPECT_STR, buf);           \
     } while (0)
 
 TEST(itoa, 0) { NPF_ITOA_CHECK("0", 0); }
